Search: convertPageToSearchPageData() конвертация в объект SearchPageData  и расчет релевантности в abs и rel в момент конвертации

+++ src/main/java/searchengine/model/IndexEntity.java
@@ -23,6 +23,6 @@ public class IndexEntity {
     private Lemma lemma;
 
     @Column(name="rating", nullable=false)
-    private Float rank;
+    private Double rank;
 
 }
